metering, audit: the counts on two result structs stop being int
Five exported result counts spelled themselves int while every other one
in the module — Deleted, Removed, Scanned, Leased, Pending, Ready, Count
and every Sweep and Reap return — answers in int64. A count's type is
exported API in the one way its value never shows: nothing a caller
writes against Flushed or Checked reads differently at either width, so
the narrowing is invisible right up until the day the field has to widen,
and widening an exported field after the tag is a /v15.

FlushResult.Claimed goes with Flushed, Skipped and Failed. The issue
named three of the four, but Claimed is the same field with the same
lifetime — it is assigned from a len, subtracted from the other two in
the backlog gauge, and sat beside Quantity and EventsReaped, which were
int64 already. Leaving it would have left the one struct spelling the
same fact two ways and the gauge converting one operand of a subtraction.

The conversions the widening removes are the evidence it was owed:
audit's gRPC converters were casting Checked in both directions because
the proto field always was an int64, and the backlog gauge was wrapping
its subtraction to reach a Record that always wanted one.

Each package pins the widths with a reflection test over its result
types, so a count added later at int fails here rather than in a
consumer's next major bump.

Closes #585

+++ audit/reader.go
@@ -189,7 +189,7 @@ type VerificationResult struct {
 	// Scope is the chain that was walked.
 	Scope tenancy.Scope
 	// Checked is how many entries were walked.
-	Checked int
+	Checked int64
 }
 
 // Intact reports whether the verified range held together. It is a method
@@ -483,7 +483,7 @@ func (r *SQLReader) Verify(ctx context.Context, scope tenancy.Scope, from, to ti
 		return nil, op.Error(err, "reading audit chain for scope %s", scope)
 	}
 
-	result := &VerificationResult{Scope: scope, From: from, To: to, Checked: len(stored)}
+	result := &VerificationResult{Scope: scope, From: from, To: to, Checked: int64(len(stored))}
 
 	if len(stored) > 0 {
 		var anchor *anchorState
